+++ CHANGELOG.md
@@ -4,6 +4,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
  * Bump runtimeVersion to `2025.08.25`.
  * Upgraded dependencies (incl. `googleapis` and `googleapis_auth`)
+ * Note: upload emails may contain a changelog excerpt.
 
 ## `20250821t095300-all`
  * Bump runtimeVersion to `2025.08.15`.